#Issue: Incorrect mimetype being set for .msg files by SharePoint connector
#Analysis
1.Connector uses the responseHeaders set by the HttpClient library for setting
the mimeType of a SPDocument.

After debugging the HttpClient library code for the issue below are my
observation:
1. The client library uses the inputStream coming from HttpConnection to set the
response headers.
2. The incoming stream has the mimeType  definition already set on the server
side.  

Hence appropriate mimeType is not received i.e(application/vnd.ms-outlook) for
.msg files.

Tried resolving the issue by adding the mimeType "application/vnd.ms-outlook" to
the WebApplication on SharePoint server but the issue still persists.

Also tried by adding the mimeType application/vnd.ms-outlook to ContentType.java
of HttpClient Library even that did not help resolve the issue.


#Reason
The HttpClient library uses the inputStream coming from HttpConnection to set
the response headers.
2. The incoming stream has the mimeType  definition already to
application/octect-stream.  

#changes
1. Modified SPDocument.java to check for file extension and set mimetype to
application/vnd.ms-outlook when it is a .msg file.

This is a crude fix for the issue with the findings(limitation) i.e is to have a
check for .msg files extension in-case of document library items and set the
mimeType as application/vnd.ms-outlook for these Items.