Adding a fix for b/5941470 (Sharepoint SOAP response has duplicate attributes). SAX throws an exception when a node in the SOAP response has duplicate attribute entries. This is caused when a document has duplicate metadata in the Sharepoint repository.
To fix this, I have to define and register a new SAX parser factory for Axis. The SAX parser factory returns a SAX parser that intercepts fatal errors and ignores the duplicate attributes exception so that parsing of the SOAP response continues to the end of the document.

When this error happens, the connector stops crawling the list. No documents get crawled for the list at all.

The workaround is for the customer to clean up the documents on Sharepoint by removing all duplicate attributes from the documents but a code change is needed here since the impact is so drastic for the customer.

Code review: http://codereview.appspot.com/5645065/