Fix bugs in part two of the SharepointConnectorType refactoring. Both were confusions about the collation and property names.
Changed the SPConstants USERNAME_FORMAT_IN_ACE and
GROUPNAME_FORMAT_IN_ACE from "Username format in ACE" to
"usernameFormatInAce", and similarly for the group name, because the
value is a fixed field key, not a display string. The collation was
apparently ignoring spaces.

Changed "Password" back to "password" everywhere. The value of
SPConstants.PASSWORD was "Password" (and is also now "password"), but
the form key in SharepointConnectorType was already "password" (and
always was, in connectorType.xml), so the property key is "password".

Code review: http://codereview.appspot.com/6569051/