11283757 Support for ADFS authentication.
It uses WS-Trust 2005 to authenticate against
the https://example/adfs/services/trust/2005/usernamemixed/ endpoint.

If we encounter 403, we do not assume claims based authentication as previously,
but check if there is runtime setting with the STS endpoint. If so, we
authenticate against the STS endpoint, retrieve the RequestSecurityTokenResponse
and bind with it against /_trust on the SharePoint server.

Will work only for offboard connectors, where the runtime setting can be modified

b/11283757
https://codereview.appspot.com/14870045/